convert: Move binpacking to png services.
