+++ openage/convert/entity_object/export/CMakeLists.txt
@@ -1,6 +1,5 @@
 add_py_modules(
 	__init__.py
-	binpack.py
 	data_definition.py
 	media_export_request.py
 	metadata_export.py
